refactor: split CLI and Chainlit modules
